fix(indexing): Correct some 16-bit indexing

+++ src/PixelBuffer.h
@@ -109,11 +109,11 @@ public:
 	uint32_t getLength() const { return length; }
 	uint32_t getSize() const { return length; }
 
-	inline T<FORMAT>& operator[] (int16_t index) __attribute__((always_inline)) {
+	inline T<FORMAT>& operator[] (int index) __attribute__((always_inline)) {
 		return pixels[index];
 	}
 
-	inline void setPixel(int16_t index, T<FORMAT> col) __attribute__((always_inline)) {
+	inline void setPixel(int index, T<FORMAT> col) __attribute__((always_inline)) {
 		pixels[index] = col;
 	}
 
@@ -132,13 +132,13 @@ public:
 		}
 	}
 
-	inline void setPixels(int16_t index, uint8_t length, T<FORMAT> col) __attribute__((always_inline)) {
+	inline void setPixels(int index, uint8_t length, T<FORMAT> col) __attribute__((always_inline)) {
 		for (auto i = index; i < index + length; i++) {
 			pixels[i] = col;
 		}
 	}
 
-	inline void setMirroredPixel(int16_t index, T<FORMAT> col) __attribute__((always_inline)) {
+	inline void setMirroredPixel(int index, T<FORMAT> col) __attribute__((always_inline)) {
 		pixels[index] = col;
 		pixels[length - index - 1] = col;
 	}
